General Code Hygene
Added two functions that help to reduce redundancy, and made some flow-of-control modifications. Added a few comments to make it a bit easier to read. No functional changes. Be gentle with me, as I don't have the ability to check for bugs, and it is, in fact, my first time.